requirements.txt
Django>=4.2
Pillow
gunicorn
whitenoise
requests
Werkzeug